refactor: 예외 처리 스타일 통일
