+++ src/main/java/kr/co/fastcampus/fastcatch/common/exception/NoCartException.java
@@ -1,0 +1,11 @@
+package kr.co.fastcampus.fastcatch.common.exception;
+
+import kr.co.fastcampus.fastcatch.common.response.ErrorCode;
+
+public class NoCartException extends BaseException {
+
+    public NoCartException() {
+        super(ErrorCode.NO_CART.getErrorMsg());
+    }
+
+}
