+++ src/main/java/kr/co/fastcampus/fastcatch/common/exception/NoCartItemException.java
@@ -1,0 +1,11 @@
+package kr.co.fastcampus.fastcatch.common.exception;
+
+import kr.co.fastcampus.fastcatch.common.response.ErrorCode;
+
+public class NoCartItemException extends BaseException {
+
+    public NoCartItemException() {
+        super(ErrorCode.NO_CART_ITEM.getErrorMsg());
+    }
+
+}
