+++ src/main/java/kr/co/fastcampus/fastcatch/common/utility/exception/StartDateIsLaterThanEndDateException.java
@@ -1,11 +1,12 @@
 package kr.co.fastcampus.fastcatch.common.utility.exception;
 
 import kr.co.fastcampus.fastcatch.common.exception.BaseException;
+import kr.co.fastcampus.fastcatch.common.response.ErrorCode;
 
 public class StartDateIsLaterThanEndDateException extends BaseException {
 
     public StartDateIsLaterThanEndDateException() {
-        super(AvailableUtilErrorCode.START_DATE_IS_LATER_THAN_END_DATE.getErrorMsg());
+        super(ErrorCode.START_DATE_IS_LATER_THAN_END_DATE.getErrorMsg());
     }
 
 }
