--- src/main/java/kr/co/fastcampus/fastcatch/domain/cart/exception/CartErrorCode.java
@@ -1,17 +1,0 @@
-package kr.co.fastcampus.fastcatch.domain.cart.exception;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum CartErrorCode {
-
-    NO_CART("해당 하는 카트가 없습니다.");
-
-    private final String errorMsg;
-
-    public String getErrorMsg(Object... arg) {
-        return String.format(errorMsg, arg);
-    }
-}
